// src/lib/generated/settings.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AppSettingsDto = { onboardingCompleted: boolean, activeProviderProfileId: string | null, defaultLearningProfileId?: string | null, defaultVisionProfileId?: string | null, theme: Theme, contextMode: ContextMode, uiLanguage: UiLanguage, uiLanguageInitialized: boolean, firstReaderHintCompleted: boolean, };

export type ContextMode = "standard" | "long";

export type Theme = "light" | "dark" | "system";

export type UiLanguage = "zh-CN" | "zh-TW" | "en";